move AnimalConfig in CoproductConverterSuite
