+++ core/src/test/scala/pureconfig/arbitrary/package.scala
@@ -16,6 +16,5 @@ package object arbitrary {
   implicit val arbUUID = Arbitrary(genUUID)
   implicit val arbPath = Arbitrary(genPath)
   implicit val arbPercentage = Arbitrary(genPercentage)
-  implicit val arbAnimalConfig = Arbitrary(genAnimalConfig)
   implicit val arbJodaDateTime = Arbitrary(genJodaDateTime)
 }
