--- core/src/test/scala/pureconfig/data/AnimalConfig.scala
@@ -1,6 +1,0 @@
-package pureconfig.data
-
-sealed trait AnimalConfig
-case class DogConfig(age: Int) extends AnimalConfig
-case class CatConfig(age: Int) extends AnimalConfig
-case class BirdConfig(canFly: Boolean) extends AnimalConfig
